+++ changelog.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add [Travis CI](https://medium.com/angular-chile/integraci%C3%B3n-continua-angular-github-firebase-travis-ci-a1a8946d471a).
 - Add tabs in UserView and profile component for audit and entities.
 
+### Changed
+
+- Change user, entity and place views.
+
 ## [0.0.8] - 2021-02-21 Evolve Cloud Functions
 
 ### Fixed